# openapi.yaml
openapi: "3.2.0"
info:
  title: sing-box hub
  version: 1.3.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  description: |
    独立的多节点 sing-box 监控管理面板。

    - **数据面**:复用每个 sing-box 节点内置的 Clash API(`experimental.clash_api`)采集
      实时流量、连接(含客户端来源 IP)与内存信息;节点侧无需安装任何组件。
    - **能力**:多节点注册与健康检查、流量历史(分钟/小时/天)、按来源 IP 的流量统计、
      出口 IP 探测、基本控制(断开连接、切换节点选择、切换 mode)、订阅导出
      (sing-box JSON / Clash YAML / base64 分享链接 URI 列表)。
    - **约定**:错误一律 RFC 9457 Problem Details(`application/problem+json`);
      列表游标分页(`cursor` + `page_size`,响应 `next_cursor` 与绝对 URI `links`);
      时间 RFC 3339 UTC;字段 snake_case;数量字段带单位后缀(`_bytes`、`_bps`、`_ms`)。
    - **订阅安全取舍**:`GET /sub/{token}` 为「能力 URL」(capability URL),令牌置于路径而非
      Authorization 头,原因是订阅客户端 App 无法携带自定义请求头。缓解:令牌可随时轮换与吊销、
      端点只读、有限流与访问日志、建议部署侧强制 HTTPS。
    - **节点不可达**:面板到节点的 Clash API 请求失败统一返回 `502`,
      `type: https://sing-hub.dev/probs/node-unreachable`。
    - **加密配置导入**:节点可选 `config_url` + `kce_key`;添加/更新时面板自动拉取
      (支持 KCE1 加密封装)并解析 Clash 代理,全部代理自动导入为面板节点;上游
      拉取/解密失败统一返回 `502`,`type: https://sing-hub.dev/probs/upstream-unavailable`。
    - **账号与会话**:所有面板账号均有管理权限。访问令牌绑定数据库中的会话;
      删除账号、修改密码或用户名、登出会撤销相关会话。刷新令牌只能消费一次。
      升级前的访问令牌需要刷新,有效的旧刷新令牌仍可换取新会话。
    - **公开状态**:`/api/public/*` 无需登录,公开节点名称、标签、出口 IP、版本和运行统计,
      不返回 API 地址、API secret、来源 IP 明细或出站配置。
servers:
  - url: http://localhost:9090
    description: 本地开发默认地址
...
security:
  - bearerAuth: []
paths:
...
  /nodes:
    get:
      operationId: listNodes
      summary: 节点列表(游标分页,含实时摘要)
      tags: [nodes]
      parameters:
        - { $ref: "#/components/parameters/cursor" }
        - { $ref: "#/components/parameters/pageSize" }
        - name: q
          in: query
          schema: { type: string }
          description: 按名称模糊过滤
        - name: tag
          in: query
          schema: { type: string }
        - name: enabled
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NodeSummaryList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createNode
      summary: 新增节点
      tags: [nodes]
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string }
          description: 幂等键;同键重放返回首次创建结果,去重窗口 24h
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NodeInput" }
            examples:
              minimal:
                value:
                  name: home-sb
                  api_url: http://192.168.1.10:9090
                  api_secret: s3cret
              full:
                value:
                  name: vps-jp
                  api_url: https://sb.example.net:9090
                  api_secret: s3cret
                  tags: [prod, jp]
                  enabled: true
                  remark: Tokyo VPS
      responses:
        "201":
          description: 已创建
          headers:
            Location:
              schema: { type: string, format: uri-reference }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
        "429": { $ref: "#/components/responses/RateLimited" }
        "500": { $ref: "#/components/responses/InternalError" }
  /nodes/{node_id}:
    parameters:
      - { $ref: "#/components/parameters/nodeId" }
    get:
      operationId: getNode
      summary: 节点详情
      tags: [nodes]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateNode
      summary: 部分更新节点(application/merge-patch+json)
      tags: [nodes]
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema: { $ref: "#/components/schemas/NodePatch" }
            examples:
              rename:
                value: { name: vps-jp-2, remark: moved }
              rotateSecret:
                value: { api_secret: new-s3cret }
      responses:
        "200":
          description: 已更新
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Node" }
        "404": { $ref: "#/components/responses/NotFound" }
        "412": { $ref: "#/components/responses/PreconditionFailed" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: deleteNode
      summary: 删除节点(连同其统计数据)
      tags: [nodes]
      responses:
        "204": { description: 已删除 }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /stats/nodes:
    get:
      operationId: getNodeDashboard
      summary: 全节点仪表盘快照
      tags: [monitoring]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items, updated_at]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/NodeDashboardItem" }
                  updated_at: { type: string, format: date-time }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalError" }
...
  /api/public/nodes:
    get:
      operationId: getPublicNodes
      summary: 公开全节点状态快照
      tags: [monitoring]
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/PublicNodeItem" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 登录 /session 获取的访问令牌
  parameters:
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: 分页游标,来自上一页 next_cursor
    pageSize:
      name: page_size
      in: query
      schema: { type: integer, maximum: 100, default: 20 }
    nodeId:
      name: node_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    InternalError:
      description: 内部错误
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    Unauthorized:
      description: 未认证
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NotFound:
      description: 资源不存在
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    Conflict:
      description: 冲突(如节点重名)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    ValidationError:
      description: 语义校验失败
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    PreconditionFailed:
      description: If-Match 不匹配
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    RateLimited:
      description: 请求过于频繁
      headers:
        Retry-After:
          schema: { type: integer }
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    NodeLiveSummary:
      type: object
      required: [id, name, tags, enabled, online, up_bps, down_bps, active_connections, today_up_bytes, today_down_bytes]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tags: { type: array, items: { type: string } }
        enabled: { type: boolean }
        online: { type: boolean }
        up_bps: { type: integer }
        down_bps: { type: integer }
        active_connections: { type: integer }
        today_up_bytes: { type: integer }
        today_down_bytes: { type: integer }
        memory_bytes: { type: [integer, "null"] }
        version: { type: [string, "null"] }
        last_online_at: { type: [string, "null"], format: date-time }
    NodeDashboardItem:
      allOf:
        - { $ref: "#/components/schemas/NodeLiveSummary" }
        - type: object
          required: [has_outbound]
          properties:
            has_outbound: { type: boolean }
            mode: { type: [string, "null"] }
    PublicNodeItem:
      allOf:
        - { $ref: "#/components/schemas/NodeLiveSummary" }
        - type: object
          required: [exit_ip]
          properties:
            exit_ip: { type: string }
    Problem:
      type: object
      description: RFC 9457 Problem Details;type 为稳定绝对 URI,未知类型为 about:blank
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        errors:
          type: array
          description: 校验错误明细,pointer 为 RFC 6901 JSON Pointer
          items:
            type: object
            required: [detail, pointer]
            properties:
              detail: { type: string }
              pointer: { type: string }
      examples:
        - type: https://sing-hub.dev/probs/node-unreachable
          title: Node unreachable
          status: 502
          detail: "dial tcp 192.168.1.10:9090: connect refused"
          instance: /nodes/3fa85f64-5717-4562-b3fc-2c963f66afa6/connections
...
    ImportStats:
      type: object
      required: [found, imported, updated, removed, skipped]
      properties:
        found: { type: integer, description: 本次解析到的代理数 }
        imported: { type: integer, description: 新导入节点数 }
        updated: { type: integer, description: 刷新出站的既有节点数 }
        removed: { type: integer, description: 上游已消失而删除的节点数 }
        skipped: { type: integer, description: 因类型不支持或字段缺失被跳过的代理数 }
    Node:
      type: object
      required: [id, name, api_url, api_secret_set, tags, enabled, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        api_url: { type: string, format: uri }
        api_secret_set:
          type: boolean
          description: 已配置密钥则为 true;密钥本身永不回读
        outbound_json:
          type: [string, "null"]
          description: 节点出站配置(sing-box outbound JSON);列表接口不返回此字段
        source:
          type: string
          description: 节点来源;''=手工添加,`sub:<id>`=由上游订阅源导入
        config_url:
          type: string
          description: 加密配置地址(KCE1 密文或明文 Clash YAML);为空表示未配置解密导入
        kce_key_set:
          type: boolean
          description: 已保存 KCE1 解密密钥则为 true;密钥本身永不回读
        import:
          description: 仅创建/更新且携带 config_url 时返回的解密导入统计
          allOf:
            - $ref: "#/components/schemas/ImportStats"
        tags:
          type: array
          items: { type: string }
        enabled: { type: boolean }
        remark: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_online_at: { type: [string, "null"], format: date-time }
    NodeInput:
      type: object
      required: [name, api_url, api_secret]
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        api_url:
          type: string
          format: uri
          description: 节点 Clash API 地址,如 http://10.0.0.2:9090
        api_secret: { type: string, minLength: 1 }
        outbound_json:
          type: [string, "null"]
          description: sing-box outbound JSON 对象文本;用于出口 IP 探测与订阅导出
        config_url:
          type: string
          format: uri
          description: >
            可选;提供时拉取该地址(支持 KCE1 密文,需配合 kce_key)并解析
            Clash 代理,出站未手动给出时取第 1 个代理的出站
        kce_key:
          type: string
          maxLength: 256
          description: 可选;config_url 返回 KCE1 密文时必需的解密密钥(保存后不回读)
        tags:
          type: array
          items: { type: string }
        enabled: { type: boolean, default: true }
        remark: { type: string }
    NodePatch:
      type: object
      description: merge-patch 语义;未提供的字段保持不变
      properties:
        name: { type: string, minLength: 1, maxLength: 64 }
        api_url: { type: string, format: uri }
        api_secret: { type: string, minLength: 1 }
        outbound_json: { type: [string, "null"] }
        config_url: { type: [string, "null"] }
        kce_key: { type: string }
        tags:
          type: array
          items: { type: string }
        enabled: { type: boolean }
        remark: { type: string }
    NodeSummary:
      type: object
      required: [id, name, tags, enabled, online, active_connections, up_bps, down_bps]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        tags:
          type: array
          items: { type: string }
        enabled: { type: boolean }
        source: { type: string, description: "''=手工,`sub:<id>`=订阅导入" }
        online: { type: boolean }
        has_outbound: { type: boolean, description: 是否配置了出站(参与订阅与出口 IP 探测) }
        up_bps: { type: integer, description: 当前上行速率 bit/s(离线为 0) }
        down_bps: { type: integer, description: 当前下行速率 bit/s }
        active_connections: { type: integer }
        last_online_at: { type: [string, "null"], format: date-time }
    NodeSummaryList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/NodeSummary" }
        next_cursor: { type: [string, "null"] }
        links:
          type: [object, "null"]
          properties:
            self: { type: string, format: uri }
            next: { type: string, format: uri }
...